Created new objects, each responsible for their calcs
